Run at document-end
Sometimes even though the dom tree is parsed not everything is done rendering. This causes event listeners to fail since they cannot get an element ID on a null/undefinted value. It seems the Chrome and Firefox do not see null/undefined the same way all the time, which made this very weird to debug